Extract default types to file
