+++ lib/model.js
@@ -4,13 +4,7 @@ const fsp = require('fs').promises;
 const { isFirstUpper } = require('metautil');
 const { Schema } = require('./schema.js');
 const { readDirectory } = require('./loader.js');
-
-const TYPES = {
-  number: 'number',
-  string: 'string',
-  boolean: 'boolean',
-  enum: 'string',
-};
+const { TYPES } = require('./types.js');
 
 class Model {
   constructor(types, entities, database = null) {
