+++ lib/types.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const TYPES = {
+  number: 'number',
+  string: 'string',
+  boolean: 'boolean',
+  enum: 'string',
+};
+
+module.exports = { TYPES };
